Refactors orchestrator code into separate files
