+++ clients/cli/src/orchestrator/mod.rs
@@ -1,0 +1,39 @@
+use crate::environment::Environment;
+use crate::orchestrator::error::OrchestratorError;
+use crate::task::Task;
+use ed25519_dalek::SigningKey;
+
+mod client;
+pub use client::OrchestratorClient;
+pub mod error;
+
+#[async_trait::async_trait]
+#[allow(dead_code)]
+pub trait Orchestrator {
+    fn environment(&self) -> &Environment;
+
+    /// Registers a new user with the orchestrator.
+    async fn register_user(
+        &self,
+        user_id: &str,
+        wallet_address: &str,
+    ) -> Result<(), OrchestratorError>;
+
+    /// Registers a new node with the orchestrator.
+    async fn register_node(&self, user_id: &str) -> Result<String, OrchestratorError>;
+
+    /// Get the list of tasks currently assigned to the node.
+    async fn get_tasks(&self, node_id: &str) -> Result<Vec<Task>, OrchestratorError>;
+
+    /// Retrieves a proof task for the node.
+    async fn get_proof_task(&self, node_id: &str) -> Result<Task, OrchestratorError>;
+
+    /// Submits a proof to the orchestrator.
+    async fn submit_proof(
+        &self,
+        task_id: &str,
+        proof_hash: &str,
+        proof: Vec<u8>,
+        signing_key: SigningKey,
+    ) -> Result<(), OrchestratorError>;
+}
